backend/requirements.txt
Django==4.2.12
django-extra-fields==3.0.2
django-filter==2.4.0
djangorestframework==3.15.2
djoser==2.1.0
drf-yasg==1.21.7
psycopg2-binary==2.9.9
Pillow==9.5.0
python-dotenv==1.0.0
gunicorn==20.1.0
environs==11.0.0